+++ app/src/main/java/com/fibelatti/photowidget/platform/BitmapKtx.kt
@@ -144,13 +144,7 @@ private fun Bitmap.sourceRect(
         bitmapWidth = width.toFloat(),
         bitmapHeight = height.toFloat(),
         aspectRatio = when (aspectRatio) {
-            PhotoWidgetAspectRatio.SQUARE, PhotoWidgetAspectRatio.TALL, PhotoWidgetAspectRatio.WIDE -> {
-                aspectRatio.aspectRatio
-            }
-
-            PhotoWidgetAspectRatio.ORIGINAL -> {
-                width / height.toFloat()
-            }
+            PhotoWidgetAspectRatio.ORIGINAL -> width / height.toFloat()
 
             PhotoWidgetAspectRatio.FILL_WIDGET -> {
                 if (widgetSize != null && widgetSize.width > 0 && widgetSize.height > 0) {
@@ -159,6 +153,8 @@ private fun Bitmap.sourceRect(
                     width / height.toFloat()
                 }
             }
+
+            else -> aspectRatio.aspectRatio
         },
     ).toRect().also { Timber.d("Output rect: $it") }
 }